requirements.txt
numpy==1.25.0
opencv-contrib-python==4.7.0.72
python-dotenv==1.0.0
pyzmq==25.1.0
